💩 to throw away: TextBox using Windows.Ui.Text.Core
While this is documented as available in Win10 RTM, it turns out it actually requires Win11 insider preview to work in unpackaged apps without corewindow.

There are rumors that it sort-of works on Win10 nowadays, but the textinput event is not raised for normal keyboard input - which removes the advantage of using it over Text Services Framework.

refs #75