⚡ 🩹 Pass a nominal 'now' through `Widget::Tick()`
We're doing *rare* frames - but also animations with state transitions, e.g. cursor blinking.

If we blink the cursor every 500ms, we need to make sure we don't skip an update because it's only been 499ms.

It would be better still to thread this through EndFrame, but that's not required. For now, I'm depending on the timers taking *at least* as long as requested.

refs #75